requirements.txt
requests==2.31.0
rich==13.4.2
python-dotenv==1.0.0
